fix warnings

+++ dsview/CableReceiver.cpp
@@ -244,7 +244,7 @@ void DSReciever::receiverThreadHandler(TRingbuffer<uint8_t>* rbuf, volatile bool
 	{
 		//cout << "Reading Block..." << endl;
 		size_t read = usb_device->Read(cable_buf.data(), cable_buf.size());
-		std::fill(cable_buf.begin() + read, cable_buf.end(), 0);
+		std::fill(cable_buf.begin() + long(read), cable_buf.end(), 0);
 		if (read != cable_buf.size())
 			cerr << "Warning: USB Read Data timeout" << endl;
 		rbuf->Put(cable_buf.data(), cable_buf.size());
